Reintroduce predict mode
